[Key Manager] Removed the Key Manager Config from Node Config and created a self-contained top-level Config.
