+++ config/src/config/secure_backend_config.rs
@@ -5,57 +5,6 @@ use anyhow::{anyhow, Result};
 use serde::{Deserialize, Serialize};
 use std::{fs::File, io::Read, path::PathBuf};
 
-// JSON RPC endpoint related defaults
-const DEFAULT_JSON_RPC_ENDPOINT: &str = "https://127.0.0.1:8080";
-
-// Key manager timing related defaults
-const DEFAULT_ROTATION_PERIOD_SECS: u64 = 604_800; // 1 week
-const DEFAULT_SLEEP_PERIOD_SECS: u64 = 600; // 10 minutes
-const DEFAULT_TXN_EXPIRATION_SECS: u64 = 3600; // 1 hour
-
-#[derive(Clone, Debug, Default, Deserialize, PartialEq, Serialize)]
-#[serde(default, deny_unknown_fields)]
-pub struct SecureConfig {
-    pub key_manager: KeyManagerConfig,
-}
-
-impl SecureConfig {
-    pub fn set_data_dir(&mut self, data_dir: PathBuf) {
-        self.key_manager.set_data_dir(data_dir);
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(default, deny_unknown_fields)]
-pub struct KeyManagerConfig {
-    pub rotation_period_secs: u64,
-    pub sleep_period_secs: u64,
-    pub txn_expiration_secs: u64,
-
-    pub json_rpc_endpoint: String,
-    pub secure_backend: SecureBackend,
-}
-
-impl Default for KeyManagerConfig {
-    fn default() -> KeyManagerConfig {
-        KeyManagerConfig {
-            rotation_period_secs: DEFAULT_ROTATION_PERIOD_SECS,
-            sleep_period_secs: DEFAULT_SLEEP_PERIOD_SECS,
-            txn_expiration_secs: DEFAULT_TXN_EXPIRATION_SECS,
-            json_rpc_endpoint: DEFAULT_JSON_RPC_ENDPOINT.into(),
-            secure_backend: SecureBackend::InMemoryStorage,
-        }
-    }
-}
-
-impl KeyManagerConfig {
-    pub fn set_data_dir(&mut self, data_dir: PathBuf) {
-        if let SecureBackend::OnDiskStorage(backend) = &mut self.secure_backend {
-            backend.set_data_dir(data_dir);
-        }
-    }
-}
-
 #[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
 #[serde(tag = "type")]
 #[serde(rename_all = "snake_case")]
