+++ secure/key-manager/src/main.rs
@@ -5,7 +5,7 @@
 
 #![forbid(unsafe_code)]
 
-use libra_config::config::{KeyManagerConfig, NetworkConfig, NodeConfig};
+use libra_config::config::KeyManagerConfig;
 use libra_key_manager::{
     counters::COUNTERS, libra_interface::JsonRpcLibraInterface, Error, KeyManager,
 };
@@ -23,39 +23,28 @@ fn main() {
         process::exit(1);
     }
 
-    let config = NodeConfig::load(&args[1]).unwrap_or_else(|e| {
+    let key_manager_config = KeyManagerConfig::load(&args[1]).unwrap_or_else(|e| {
         eprintln!(
-            "Error! Unable to load provided config: {}, error: {}",
+            "Error! Unable to load provided key manager config: {}, error: {}",
             args[1], e
         );
         process::exit(1);
     });
 
-    if config.validator_network.is_none() {
-        eprintln!("Error! Validator config missing from node config.");
-        process::exit(1);
-    }
-    let network_config = config.validator_network.unwrap();
-    let key_manager_config = config.secure.key_manager;
-
     libra_logger::Logger::new()
-        .channel_size(config.logger.chan_size)
-        .is_async(config.logger.is_async)
-        .level(config.logger.level)
+        .channel_size(key_manager_config.logger.chan_size)
+        .is_async(key_manager_config.logger.is_async)
+        .level(key_manager_config.logger.level)
         .init();
     MetricsPusher::new(COUNTERS.clone()).start();
 
-    create_and_execute_key_manager(network_config, key_manager_config).unwrap_or_else(|e| {
+    create_and_execute_key_manager(key_manager_config).unwrap_or_else(|e| {
         eprintln!("Error! The Key Manager has failed during execution: {}", e);
         process::exit(1);
     });
 }
 
-fn create_and_execute_key_manager(
-    network_config: NetworkConfig,
-    key_manager_config: KeyManagerConfig,
-) -> Result<(), Error> {
-    let account = network_config.peer_id;
+fn create_and_execute_key_manager(key_manager_config: KeyManagerConfig) -> Result<(), Error> {
     let libra_interface = create_libra_interface(key_manager_config.json_rpc_endpoint);
     let storage: Box<dyn Storage> = (&key_manager_config.secure_backend)
         .try_into()
@@ -63,7 +52,7 @@ fn create_and_execute_key_manager(
     let time_service = RealTimeService::new();
 
     KeyManager::new(
-        account,
+        key_manager_config.validator_account,
         libra_interface,
         BoxStorage(storage),
         time_service,
